+++ spec/lib/aixm/component/fato_spec.rb
@@ -1,8 +1,8 @@
 require_relative '../../../spec_helper'
 
-describe AIXM::Component::Helipad do
+describe AIXM::Component::FATO do
   subject do
-    AIXM::Factory.airport.helipads.first
+    AIXM::Factory.airport.fatos.first
   end
 
   describe :name= do
@@ -13,18 +13,6 @@ describe AIXM::Component::Helipad do
     it "upcases and transcodes valid values" do
       subject.tap { |s| s.name = 'h1' }.name.must_equal 'H1'
     end
-  end
-
-  describe :xy= do
-    macro :xy
-
-    it "fails on nil value" do
-      [nil].wont_be_written_to subject, :xy
-    end
-  end
-
-  describe :z= do
-    macro :z_qnh
   end
 
   describe :length= do
@@ -47,20 +35,17 @@ describe AIXM::Component::Helipad do
     end
   end
 
-  describe :helicopter_class= do
-    it "fails on invalid values" do
-      [:foobar, 123].wont_be_written_to subject, :helicopter_class
-    end
-
+  describe :profile= do
     it "accepts nil value" do
-      [nil].must_be_written_to subject, :helicopter_class
+      [nil].must_be_written_to subject, :profile
     end
 
-    it "looks up valid values" do
-      subject.tap { |s| s.helicopter_class = 1 }.helicopter_class.must_equal :'1'
-      subject.tap { |s| s.helicopter_class = :OTHER }.helicopter_class.must_equal :other
+    it "stringifies valid values" do
+      subject.tap { |s| s.profile = 'foobar' }.profile.must_equal 'foobar'
+      subject.tap { |s| s.profile = 123 }.profile.must_equal '123'
     end
   end
+
 
   describe :marking= do
     macro :marking
@@ -89,26 +74,15 @@ describe AIXM::Component::Helipad do
     it "builds correct complete OFMX" do
       AIXM.ofmx!
       subject.to_xml.must_equal <<~END
-        <Tla>
-          <TlaUid>
-            <AhpUid>
-              <codeId>LFNT</codeId>
-            </AhpUid>
-            <txtDesig>H1</txtDesig>
-          </TlaUid>
+        <Fto>
           <FtoUid>
             <AhpUid>
               <codeId>LFNT</codeId>
             </AhpUid>
             <txtDesig>H1</txtDesig>
           </FtoUid>
-          <geoLat>43.99915000N</geoLat>
-          <geoLong>004.75154444E</geoLong>
-          <codeDatum>WGE</codeDatum>
-          <valElev>141</valElev>
-          <uomDistVer>FT</uomDistVer>
-          <valLen>20</valLen>
-          <valWid>20</valWid>
+          <valLen>35</valLen>
+          <valWid>35</valWid>
           <uomDim>M</uomDim>
           <codeComposition>CONC</codeComposition>
           <codePreparation>PAVED</codePreparation>
@@ -119,36 +93,27 @@ describe AIXM::Component::Helipad do
           <codePcnMaxTirePressure>W</codePcnMaxTirePressure>
           <codePcnEvalMethod>U</codePcnEvalMethod>
           <txtPcnNote>Cracks near the center</txtPcnNote>
-          <codeClassHel>1</codeClassHel>
-          <txtMarking>Continuous white lines</txtMarking>
+          <txtProfile>Northwest from RWY 12/30</txtProfile>
+          <txtMarking>Dashed white lines</txtMarking>
           <codeSts>OTHER</codeSts>
           <txtRmk>Authorizaton by AD operator required</txtRmk>
-        </Tla>
+        </Fto>
       END
     end
 
     it "builds correct minimal OFMX" do
       AIXM.ofmx!
-      %i(z length width helicopter_class marking status remarks).each { |a| subject.send(:"#{a}=", nil) }
+      %i(length width profile marking status remarks).each { |a| subject.send(:"#{a}=", nil) }
       %i(composition preparation condition pcn remarks).each { |a| subject.surface.send(:"#{a}=", nil) }
       subject.to_xml.must_equal <<~END
-        <Tla>
-          <TlaUid>
-            <AhpUid>
-              <codeId>LFNT</codeId>
-            </AhpUid>
-            <txtDesig>H1</txtDesig>
-          </TlaUid>
+        <Fto>
           <FtoUid>
             <AhpUid>
               <codeId>LFNT</codeId>
             </AhpUid>
             <txtDesig>H1</txtDesig>
           </FtoUid>
-          <geoLat>43.99915000N</geoLat>
-          <geoLong>004.75154444E</geoLong>
-          <codeDatum>WGE</codeDatum>
-        </Tla>
+        </Fto>
       END
     end
   end