+++ spec/lib/aixm/feature/airport_spec.rb
@@ -319,6 +319,30 @@ describe AIXM::Feature::Airport do
           <uomDist>M</uomDist>
           <txtRmk>back remarks</txtRmk>
         </Rdd>
+        <Fto>
+          <FtoUid>
+            <AhpUid>
+              <codeId>LFNT</codeId>
+            </AhpUid>
+            <txtDesig>H1</txtDesig>
+          </FtoUid>
+          <valLen>35</valLen>
+          <valWid>35</valWid>
+          <uomDim>M</uomDim>
+          <codeComposition>CONC</codeComposition>
+          <codePreparation>PAVED</codePreparation>
+          <codeCondSfc>FAIR</codeCondSfc>
+          <valPcnClass>30</valPcnClass>
+          <codePcnPavementType>F</codePcnPavementType>
+          <codePcnPavementSubgrade>A</codePcnPavementSubgrade>
+          <codePcnMaxTirePressure>W</codePcnMaxTirePressure>
+          <codePcnEvalMethod>U</codePcnEvalMethod>
+          <txtPcnNote>Cracks near the center</txtPcnNote>
+          <txtProfile>Northwest from RWY 12/30</txtProfile>
+          <txtMarking>Dashed white lines</txtMarking>
+          <codeSts>OTHER</codeSts>
+          <txtRmk>Authorizaton by AD operator required</txtRmk>
+        </Fto>
         <Tla>
           <TlaUid>
             <AhpUid>
@@ -326,6 +350,12 @@ describe AIXM::Feature::Airport do
             </AhpUid>
             <txtDesig>H1</txtDesig>
           </TlaUid>
+          <FtoUid>
+            <AhpUid>
+              <codeId>LFNT</codeId>
+            </AhpUid>
+            <txtDesig>H1</txtDesig>
+          </FtoUid>
           <geoLat>43.99915000N</geoLat>
           <geoLong>004.75154444E</geoLong>
           <codeDatum>WGE</codeDatum>
@@ -342,7 +372,9 @@ describe AIXM::Feature::Airport do
           <codePcnPavementSubgrade>A</codePcnPavementSubgrade>
           <codePcnMaxTirePressure>W</codePcnMaxTirePressure>
           <codePcnEvalMethod>U</codePcnEvalMethod>
-          <txtPcnNote>Cracks near the center.</txtPcnNote>
+          <txtPcnNote>Cracks near the center</txtPcnNote>
+          <codeClassHel>1</codeClassHel>
+          <txtMarking>Continuous white lines</txtMarking>
           <codeSts>OTHER</codeSts>
           <txtRmk>Authorizaton by AD operator required</txtRmk>
         </Tla>
@@ -415,6 +447,7 @@ describe AIXM::Feature::Airport do
       subject.z = subject.declination = subject.transition_z = subject.remarks = nil
       subject.instance_variable_set(:'@addresses', [])
       subject.instance_variable_set(:'@runways', [])
+      subject.instance_variable_set(:'@fatos', [])
       subject.instance_variable_set(:'@helipads', [])
       subject.instance_variable_set(:'@usage_limitations', [])
       subject.to_xml.must_equal <<~END